add js error reporting
this currently only works with Blink-based browsers since Firefox
doesn't implement global error handlers for SDK addons.

this also introduces a random user id which, when provided by a
user reporting a bug, can be used to find their error reports.
the user id is visible on the preferences page (along with some
info on where to report bugs.)