add full i18n support
this introduces a full i18n library (i18next) which can be used to
translate OB UI elements as well as game strings OB works with.

translations are managed in .json files in locales/ (one file per
language). english and dutch are the first two supported languages,
adding others is as simple as creating a new .json file and adding
the @resource reference in the meta block. in case translations
are missing or the language detection fails, en will be used as
fallback.

previously, language detection was done by checking the TLD. This,
however, didn't work when users manually switched to non-default
languages (e.g. dutch on .com). Instead, OB is now looking for a
known string in the menu to detect the language.

this is still very much a work-in-progress as most strings still
need to be extracted and translated.