add analytics beacon including update checker
A beacon is sent to the API on load, including an anonymous user
identifier, the OB version and the game domain. This allows us to
keep track of update deployments and usage numbers on a per-domain
level.

The API endpoint returns the release notes of the latest version
if a new version is available. This is based on the releases
listed on GitHub. An update prompt is shown once every 24 hours,
encouraging users to update. Skipping versions is supported.