Fix for OUJS
Hey all,

OUJS has made a change recently to require SPDX codes for OSI approved licensing.

In order to improve the appearance of your script homepages it would be appreciated if you could modify all of your affected scripts. It's the same code in the upper right hand corner of GH here. :)

The SPDX short identifier for `MIT` is that... a few of your other scripts have `The MIT License (MIT)`.

Until this change is made you will be unable to update those affected scripts.

Thanks,
OUJS Staff